Crate bevy_mod_picking

Structs

CustomHighlightablePickingPlugin

DebugCursorPickingPlugin

DebugEventsPickingPlugin

DefaultPickingPlugins

HighlightablePickingPlugin

Hover

Tracks the current hover state to be used with change tracking in the events system.

InteractablePickingPlugin

MeshButtonMaterials

NoDeselect

Marker struct used to mark pickable entities for which you don’t want to trigger a deselection event when picked. This is useful for gizmos or other pickable UI entities.

PausedForBlockers

PickableBundle

PickableButton

A default highlightable button material implementation of the [IsPickableButton] trait that uses bevy’s StandardMaterial for highlighting meshes. You may want to implement your own component if the pickable object being rendered is not a mesh, or doesn’t use the StandardMaterial component for rendered appearance.

PickingBlocker

Marker component for entities that, whenever their Interaction component is anything other than None, will suspend highlighting and selecting PickableMeshs. Bevy UI Nodes have this behavior by default.

PickingCameraBundle

PickingPlugin

PickingPluginsState

PickingRaycastSet

This unit struct is used to tag the generic ray casting types RayCastMesh and RayCastSource. This means that all Picking ray casts are of the same type. Consequently, any meshes or ray sources that are being used by the picking plugin can be used by other ray casting systems because they will have distinct types, e.g.: RayCastMesh<PickingRaycastSet> vs. RayCastMesh<MySuperCoolRaycastingType>, and as such wil not result in collisions.

RayCastSource

The RayCastSource component is used to generate rays with the specified cast_method. A ray is generated when the RayCastSource is initialized, either by waiting for update_raycast system to process the ray, or by using a with_ray function.

Selection

Tracks the current selection state to be used with change tracking in the events system. Meshes with Selection will have selection state managed.

StandardMaterialPickingColors

Enums

HoverEvent

An event that triggers when the hover state of a Hover enabled PickableMesh changes.

PickingEvent

An event that wraps selection and hover events

PickingSystem

Primitive3d

SelectionEvent

An event that triggers when the selection state of a Selection enabled PickableMesh changes.

UpdatePicks

Traits

FromWorldHelper

Functions

event_debug_system

Listens for HoverEvent and SelectionEvent events and prints them

get_initial_mesh_button_material

mesh_events_system

Looks for changes in selection or hover state, and sends the appropriate events

mesh_focus

mesh_highlighting

mesh_selection

pause_for_picking_blockers

update_pick_source_positions

Update Screenspace ray cast sources with the current mouse position

Type Definitions

PickableMesh

A type alias for the concrete RayCastMesh type used for Picking.

PickingCamera

A type alias for the concrete RayCastSource type used for Picking.